+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Unreleased
 ----------
 ### Added
 * Add en-AU to supported translations list
+* Add enzyme i18n helpers: shallowWithIntl, mountWithIntl, renderWithIntl, and mockIntl
 
 4.23.1 - (February 5, 2019)
 ----------