+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Unreleased
 ### Added
 * Add en-AU to supported translations list
 * Add enzyme i18n helpers: shallowWithIntl, mountWithIntl, renderWithIntl, and mockIntl
+* Add doc on enzyme intl helpers, update Readme
 
 4.23.1 - (February 5, 2019)
 ----------