+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Unreleased
 ----------
 ### Added
 * Added a new method `matchScreenshotAndTestStandards` to both do a screenshot comparison and test accessibility
+* Add en-AU to supported translations list
 
 4.23.1 - (February 5, 2019)
 ----------