+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Changelog
 Unreleased
 ----------
 ### Added
-* Added a new method `matchScreenshotAndTestStandards` to both do a screenshot comparison and test accessibility
+* Added a new method `validateElement` to both do a screenshot comparison and test accessibility
 * Add en-AU to supported translations list
 
 4.23.1 - (February 5, 2019)