+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Unreleased
 * Updated aggregate translations script fallback to base locale if translation is missing for regional locale on a string-by-string basis.
 
 ### Added
+* Added a script to aggregate dependency themes into a single file
 * Add en-AU to supported translations list
 
 4.23.1 - (February 5, 2019)